feat: Add protect Token

+++ server/src/controllers/books.controllers.js
@@ -1,26 +1,30 @@
 import { Book } from "../models/book.model.js";
+import { User } from "../models/user.model.js";
 
 export const getAllBooks = async (req, res) => {
   try {
+
+
     var filter = {}
-    const {author, editorial, genre, search} = req.query
+    const { author, editorial, genre, search } = req.query
 
-    if(search){
-      filter = {...filter, title:search}
+    if (search) {
+      filter = { ...filter, title: search }
     }
-    else{
-      if(author){
-            filter = {...filter, author:author}
-          }
-
-      if(editorial){
-        filter = {...filter, editorial:editorial}
+    else {
+      if (author) {
+        filter = { ...filter, author: author }
       }
 
-      if(genre){
-        filter = {...filter, genre:genre}
+      if (editorial) {
+        filter = { ...filter, editorial: editorial }
+      }
+
+      if (genre) {
+        filter = { ...filter, genre: genre }
       }
     }
+
 
     const regexFilter = {};
     for (const key in filter) {
@@ -30,7 +34,8 @@ export const getAllBooks = async (req, res) => {
     const allBooks = await Book.find({});
     const filteredBooks = await Book.find(regexFilter)
 
-    res.status(201).json({allBooks: allBooks, filteredBooks:filteredBooks});
+    res.status(201).json({ allBooks: allBooks, filteredBooks: filteredBooks });
+
   } catch (error) {
     res.status(400).send(error.message);
   }
@@ -38,6 +43,19 @@ export const getAllBooks = async (req, res) => {
 
 export const postBook = async (req, res) => {
   try {
+    const user = req.user
+
+    const userFound = await User.findById(user.uid)
+
+    if (userFound.role !== 'admin') {
+
+      return res.status(401).json({
+        ok: false,
+        msg: 'No tiene privilegios para realizar esta accion'
+      })
+    }
+
+
     const data = req.body;
     const newBook = new Book(data);
 
@@ -50,6 +68,19 @@ export const postBook = async (req, res) => {
 
 export const updateBook = async (req, res) => {
   try {
+
+    const user = req.user
+
+    const userFound = await User.findById(user.uid)
+
+    if (userFound.role !== 'admin') {
+
+      return res.status(401).json({
+        ok: false,
+        msg: 'No tiene privilegios para realizar esta accion'
+      })
+    }
+
     const id = req.params.id;
     const data = req.body;
 
@@ -59,7 +90,7 @@ export const updateBook = async (req, res) => {
         $set: data,
       }
     );
-    res.status(201).send("The book has been successfully modified");
+    res.status(201).send("EL libro ha sido modificado con exito");
   } catch (error) {
     res.status(400).send(error.message);
   }
@@ -71,16 +102,30 @@ export const getBookById = async (req, res) => {
     const result = await Book.find({ _id: id });
     res.status(200).json(result);
   } catch (error) {
-    res.status(400).send(`There is no book with that ID`);
+    res.status(400).send(`No existe un libro con ese ID`);
   }
 };
 
 export const deleteBook = async (req, res) => {
   try {
+
+    const user = req.user
+
+    const userFound = await User.findById(user.uid)
+
+    if (userFound.role !== 'admin') {
+
+      return res.status(401).json({
+        ok: false,
+        msg: 'No tiene privilegios para realizar esta accion'
+      })
+    }
+
     const id = req.params.id;
+
     await Book.deleteOne({ _id: id });
-    res.status(200).send("The book was successfully deleted");
+    res.status(200).send("El libro fue eliminado con exito");
   } catch (error) {
-    res.status(400).send(`There is no book with that ID`);
+    res.status(400).send(`No existe un libro con ese ID`);
   }
 };
